JMLL-1260 Update HR number fields to allow values in intervals of .25
* Update HR number fields to allow decimals of .25, .5, .75

* Add help text

* Use a cleaner condition

+++ leavers/models.py
@@ -152,14 +152,24 @@ class LeavingRequest(models.Model):
         blank=True,
         null=True,
     )
-    annual_number = models.IntegerField(null=True, blank=True)
+    annual_number = models.DecimalField(
+        decimal_places=2,
+        max_digits=7,
+        null=True,
+        blank=True,
+    )
     flexi_leave = models.CharField(
         choices=FlexiLeavePaidOrDeducted.choices,
         max_length=255,
         blank=True,
         null=True,
     )
-    flexi_number = models.IntegerField(null=True, blank=True)
+    flexi_number = models.DecimalField(
+        decimal_places=2,
+        max_digits=7,
+        null=True,
+        blank=True,
+    )
 
     line_reports = models.JSONField(null=True, blank=True)
